Keep using <pre> for code blocks.
