+++ src/components/code/_code_block.scss
@@ -50,7 +50,7 @@
     padding:  0 $euiSizeXS;
     background: $euiColorLightestShade;
 
-    .euiCodeBlock__pre, .euiCodeBlock__code {
+    .euiCodeBlock__code {
       display: inline;
       white-space: normal;
     }
